Adds controllers
